fix: add fallback to snapshot generator to check previous day
Due to timezones, it's possible that market snapshots will be empty for the current day. This will add a check and repoll of market snapshot of all tickers that returns empty for the previous day for that ticker.
NOTE: Not sure if there will be an issue if the previous day also returns empty. Be on the lookout for this.