feat: add new basic alert thresholds
As discussed in issue #4, issue thresholds for basic alerts are added at +-10%, +-12%, and +-14% price changes.
The initial alert at +-6% as well was updated to +-7%.
This commit encompasses multiple files since the logic structure of checking and calling multiple alerts requires an updated conditional structure.
This is incomplete to the requirements of #4, so there will still be more commits to this branch afterwards.