feat: add two new config variables for alert retriggers
This is in preparation for the next commit, which will be augmenting the alert system to also handle retriggers and updating.
The new variables, `ENABLE_ALERT_RETRIGGERING` and `ALERT_RETRIGGER_COOLDOWN_SECONDS` are set as boolean and float, respectively.
These are to indicate if alerts are allowed to retrigger given the time constraint, and to modify said time constraint.