feat: enable basic alert retrigger mechanism
Given the new configuration values from the previous commit, this now allows for a (configurable)
consecutive retriggering of alerts.
This works by checking if a basic alert has already been fired after a cooldown period (initially
set to 60 seconds). If so, then the alert is refired.
This also will signify if any alerts are actually retriggers or not, for better visuals.
NOTE: I do not trust this is exactly what was asked for, or if this was overengineered as a
solution. This might need further review.