feat: update position details of the alert message body
As mentioned in #9, this is to change the format of the `position_lines` section of the position details of the alert.
More specifically, this breaks up the structure to be separated per line, and to include OTM and trade price details
per position.
Changes done must be minimal, only affecting two files. This should proceed properly.