feat: add summary counts for positions of that ticker in alert messages
Basically, all alert messages will now also state the total number of calls and puts per ticker in that snapshot.
The summary counts will be appended at the end of the positions details section of the alert.
This also updates the Slack and Telegram service pipelines to accommodate this change.