fix: apply a filter for checking PositionSummary objects before alert firing
This is an attempt to solve the problem raised in #9, where an alert with a position with positive QTY
was wrongly fired.
Best guess as to what the issue might be is that when a blank PositionSummary object would be created
if the provided position would've resulted in no alerts. This is probably why an alert fired with the
message "No position details available".
This should now filter out any PositionSummary objects that are empty before passing them for alert
generation.