fix: patch directional check logic for firing positional alerts
This is an attempt to fix the other issue mentioned in #9. It seemed like the alert checker was ignoring the direction checks.
So when a ticker had never been alerted before, an alert would fire regardless of direction.
Also, for retrigger and incremental alerts, direction checks were also ignored.
This commit should hopefully emsure that CALLs only fire when stock is up, and PUTs only fire when stock goes down.
Additionally, the direction checks should be done for all alert types.